apps/api/requirements.txt
django>=5.0,<6.0
djangorestframework
django-cors-headers
psycopg[binary]
python-dotenv
h3
redis
celery
feedparser
beautifulsoup4
requests
pyyaml
drf-spectacular
pandas
numpy
